+++ test-harness/src/consts.ts
@@ -1,0 +1,3 @@
+import * as process from 'process';
+
+export const IS_WINDOWS = process.platform === 'win32';